+++ src/date_time/display.rs
@@ -7,7 +7,6 @@ impl fmt::Display for DateTime {
             Self::DateAndTime(dt) => write!(f, "{dt}"),
             Self::Time(t) => write!(f, "{t}"),
             Self::Date(d) => write!(f, "{d}"),
-            Self::NaiveDateAndTime(dt) => write!(f, "{dt}"),
         }
     }
 }