+++ src/options/mod.rs
@@ -18,7 +18,6 @@ pub struct Options {
     pub offset: (u32, u32),
     pub overwrite_values: bool,
     pub sheet_name: String,
-    pub time_zone: Option<chrono_tz::Tz>,
     pub use_cache: bool,
     pub verbosity: log::LevelFilter,
 }