+++ docs/CHANGELOG.md
@@ -3,6 +3,11 @@
 ### Features
 
 * Eval rows in parallel
+* Dates, times and datetimes are now tz-unaware
+
+### **Breaking Changes**
+
+* DateTime no longer supports a variant with fixed TZ offset
 
 ## v0.5.1